+++ firestore-bigquery-export/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 feat - migrate to v2 functions and support non-default firestore instances
 
+fix - fix enqueue logic and types
+
 ## Version 0.1.60
 
 feat - configure a log level to control the verbosity of logs.